Replace original coyote with Emma's updated version
